backend/app-main/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.13.0
aiosignal==1.4.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.1
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
cryptography==44.0.3
Deprecated==1.2.14
Django==4.2.25
django-cas-ng==4.3.0
django-extensions==3.2.3
djangorestframework==3.15.2
django-redis==5.4.0
drf-yasg==1.21.7
gunicorn==23.0.0
whitenoise==6.6.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.1.0
idna==3.7
importlib-metadata==6.11.0
inflection==0.5.1
ldap3==2.9.1
lxml==5.1.0
microsoft-kiota-abstractions==1.3.1
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.1
msal==1.27.0
msal-extensions==1.1.0
msgraph-core==1.0.0
psycopg2-binary==2.9.9
multidict==6.0.5
mysqlclient==2.2.4
numpy==2.1.2
openai==0.27.10
openpyxl==3.1.5
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
packaging==24.0
pandas==2.2.3
ping3==4.0.5
portalocker==2.8.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.10.1
PyMySQL==1.1.1
pyodbc==5.1.0
python-cas==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.5
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
realtime==2.22.0
supabase==2.22.0
std-uritemplate==0.0.55
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.5.0
wrapt==1.16.0
yarl==1.22.0
zipp==3.19.1
